fix(rt): hold camera distortion flag as uint32 so the intrinsics structs are padding-free
PinholeCamera / SensorMountRow are uploaded + memcmp'd byte-for-byte (the device-
vs-host camera-compose gate). A uint8 distortion among floats left 3 padding bytes
that diverged between the device kernel local and the host reference. Widen the
flag to 4 bytes (the schema CameraIntrinsics.distortion stays uint8).

[skip ci]

+++ src/phi/backend_cuda/rt/sensor_scatter.hpp
@@ -65,7 +65,7 @@ struct SensorMountRow {
     float cy = 0.0f;
     float k1 = 0.0f;
     float k2 = 0.0f;
-    uint8_t distortion = 0u;
+    uint32_t distortion = 0u;  // 4-byte so the row is padding-free (byte-exact memcmp)
     float near_clip = 0.01f;
     float far_clip = 1000.0f;
 };
